+++ flavors/vanilla/default.nix
@@ -8,7 +8,7 @@ let
 
   phoneDeviceFamilies =
     (optional (config.androidVersion <= 10) "marlin")
-    ++ [ "taimen" "muskie" "crosshatch" "bonito" "coral" "sunfish" ];
+    ++ [ "taimen" "muskie" "crosshatch" "bonito" "coral" "sunfish" "redfin" ];
   supportedDeviceFamilies = phoneDeviceFamilies ++ [ "generic" ];
 
   # Replaces references to SystemUIGoogle with SystemUI in device source tree
@@ -135,19 +135,21 @@ in mkIf (config.flavor == "vanilla") (mkMerge [
 ### Android 11 stuff ###
 (mkIf (config.androidVersion == 11) (mkMerge [
 {
-  buildDateTime = mkDefault 1612285669;
+  buildDateTime = mkDefault 1617657301;
 
-  source.manifest.rev = mkDefault "android-11.0.0_r29";
-  apv.buildID = mkDefault "RQ1A.210205.004";
+  source.manifest.rev = mkDefault "android-11.0.0_r34";
+  apv.buildID = mkDefault "RQ2A.210405.005";
 
   # See also: https://github.com/GrapheneOS/os_issue_tracker/issues/325
   # List of biometric sensors on the device, in decreasing strength. Consumed by AuthService
   # when registering authenticators with BiometricService. Format must be ID:Modality:Strength,
   # where: IDs are unique per device, Modality as defined in BiometricAuthenticator.java,
   # and Strength as defined in Authenticators.java
+  # TODO: This ought to show up in the vendor (not system or product) resource overlay
   resources."frameworks/base/core/res".config_biometric_sensors = {
-    value = optional (elem config.deviceFamily [ "taimen" "muskie" "crosshatch" "bonito" ]) "0:2:15"
-            ++ optional (config.deviceFamily == "coral") "0:8:15";
+    value = optional (elem config.deviceFamily phoneDeviceFamilies) (
+              if (config.deviceFamily == "coral") then "0:8:15"
+              else "0:2:15");
     type = "string-array";
   };
 
@@ -159,18 +161,7 @@ in mkIf (config.flavor == "vanilla") (mkMerge [
     })
   ];
 }
-(mkIf (elem config.deviceFamily phoneDeviceFamilies) (let
-  kernelTag = {
-    "taimen" = "android-11.0.0_r0.34";
-    "muskie" = "android-11.0.0_r0.34";
-    "crosshatch" = "android-11.0.0_r0.52";
-    "bonito" = "android-11.0.0_r0.53";
-    "coral" = "android-11.0.0_r0.54";
-    "sunfish" = "android-11.0.0_r0.55";
-  }.${config.deviceFamily};
-  kernelMetadata = (lib.importJSON ./kernel-metadata.json).${kernelTag};
-  kernelHashes = (lib.importJSON ./kernel-hashes.json).${kernelTag};
-in {
+(mkIf (elem config.deviceFamily phoneDeviceFamilies) {
   kernel.enable = mkDefault true;
   kernel.configName = mkMerge [
     (mkIf (elem config.deviceFamily [ "taimen" "muskie" ]) "wahoo")
@@ -181,18 +172,45 @@ in {
   # that works for kernels too. Probably not worth the effort for the payoff
   # though.
   kernel.src = let
-    fetchedRepo = repo: pkgs.fetchgit {
-      inherit (kernelHashes."https://android.googlesource.com/${repo}") url rev sha256;
+    kernelName = if elem config.deviceFamily [ "taimen" "muskie"] then "wahoo" else config.deviceFamily;
+    kernelMetadata = (lib.importJSON ./kernel-metadata.json).${kernelName};
+    kernelRepos = lib.importJSON (./. + "/repo-${kernelMetadata.branch}.json");
+    fetchRepo = repo: pkgs.fetchgit {
+      inherit (kernelRepos.${repo}) url rev sha256;
+    };
+    kernelDirs = {
+      "" = fetchRepo "private/msm-google";
+    } // optionalAttrs (elem kernelName [ "crosshatch" "bonito" "coral" "sunfish" "redfin" ]) {
+      "techpack/audio" = fetchRepo "private/msm-google/techpack/audio";
+      "drivers/staging/qca-wifi-host-cmn" = fetchRepo "private/msm-google-modules/wlan/qca-wifi-host-cmn";
+      "drivers/staging/qcacld-3.0" = fetchRepo "private/msm-google-modules/wlan/qcacld-3.0";
+      "drivers/staging/fw-api" = fetchRepo "private/msm-google-modules/wlan/fw-api";
+    } // optionalAttrs (elem kernelName [ "coral" "sunfish" ]) {
+      # Sunfish previously used a fts_touch_s5 repo, but it's tag moved back to
+      # to regular fts_touch repo, however, the kernel manifest was not updated.
+      "drivers/input/touchscreen/fts_touch" = fetchRepo "private/msm-google-modules/touch/fts/floral";
+    } // optionalAttrs (elem kernelName [ "redfin" ]) {
+      "drivers/input/touchscreen/fts_touch" = fetchRepo "private/msm-google-modules/touch/fts";
+
+      "techpack/audio" = fetchRepo "private/msm-google/techpack/audio";
+      "techpack/camera" = fetchRepo "private/msm-google/techpack/camera";
+      "techpack/dataipa" = fetchRepo "private/msm-google/techpack/dataipa";
+      "techpack/display" = fetchRepo "private/msm-google/techpack/display";
+      "techpack/video" = fetchRepo "private/msm-google/techpack/video";
+      "drivers/input/touchscreen/sec_touch" = fetchRepo "private/msm-google-modules/touch/sec";
+      "arch/arm64/boot/dts/vendor" = fetchRepo "private/msm-google/arch/arm64/boot/dts/vendor";
+      "arch/arm64/boot/dts/vendor/qcom/camera" = fetchRepo "private/msm-google/arch/arm64/boot/dts/vendor/qcom/camera";
+      "arch/arm64/boot/dts/vendor/qcom/display" = fetchRepo "private/msm-google/arch/arm64/boot/dts/vendor/qcom/display";
     };
   in pkgs.runCommand "kernel-src" {}
-    (concatStringsSep "\n" (lib.mapAttrsToList (repo: relpath: ''
+    (concatStringsSep "\n" (lib.mapAttrsToList (relpath: repo: ''
       ${lib.optionalString (relpath != "") "mkdir -p $out/$(dirname ${relpath})"}
-      cp -r ${fetchedRepo repo} $out/${relpath}
+      cp -r ${repo} $out/${relpath}
       chmod u+w -R $out/${relpath}
-    '') kernelMetadata));
+    '') kernelDirs));
 
   kernel.buildProductFilenames = [ "**/*.ko" ]; # Copy kernel modules if they exist
-}))
+})
 (mkIf (elem config.device [ "taimen" "walleye" ]) {
   warnings = [ "taimen and walleye are no longer receiving monthly vendor security updates from Google. Support is left just for testing" ];
   source.manifest.rev = "android-11.0.0_r25"; # More recent sources don't even include device/google/muskie