+++ mk-repo-file.py
@@ -51,7 +51,8 @@ def ls_remote(url, rev):
 def make_repo_file(url: str, ref: str, filename: str, ref_type: ManifestRefType,
                    override_project_revs: Dict[str, str], force_refresh: bool,
                    mirrors: Dict[str, str], project_fetch_submodules: List[str],
-                   include_prefix: List[str]):
+                   override_tag: Optional[str], include_prefix: List[str],
+                   exclude_path: List[str]):
     if os.path.exists(filename) and not force_refresh:
         data = json.load(open(filename))
     else:
@@ -67,12 +68,18 @@ def make_repo_file(url: str, ref: str, filename: str, ref_type: ManifestRefType,
         if len(include_prefix) > 0 and (not any(relpath.startswith(p) for p in include_prefix)):
             continue
 
+        if relpath in exclude_path:
+            continue
+
         for project, rev in override_project_revs.items():
             # We have to iterate over the whole output since we don't save
             # the project name anymore, just the relpath, which isn't
             # exactly the project name
             if p['url'].endswith(project):
                 p['rev'] = rev
+
+        if override_tag is not None:
+            p['revisionExpr'] = override_tag
 
         if 'rev' not in p:
             if re.match("[0-9a-f]{40}", p['revisionExpr']):
@@ -110,8 +117,12 @@ def make_repo_file(url: str, ref: str, filename: str, ref_type: ManifestRefType,
             if found_treehash:
                 continue
 
-            # Fetch information
-            git_info = checkout_git(p_url, p['rev'], fetch_submodules)
+            # Fetch information. Use revisionExpr if it is a tag so we use the
+            # tag in the name of the nix derivation instead of the revision
+            if p['revisionExpr'].startswith('refs/tags/'):
+                git_info = checkout_git(p_url, p['revisionExpr'], fetch_submodules)
+            else:
+                git_info = checkout_git(p_url, p['rev'], fetch_submodules)
             p['sha256'] = git_info['sha256']
 
             # Add to cache
@@ -132,8 +143,10 @@ def main():
                         choices=[t.name.lower() for t in ManifestRefType], default=ManifestRefType.TAG.name.lower())
     parser.add_argument('--force', help="force a re-download. Useful with --ref-type branch", action='store_true')
     parser.add_argument('--repo-prop', help="repo.prop file to use as source for project git revisions")
+    parser.add_argument('--override-tag', help="tag to fetch for subrepos, ignoring revisions from manifest")
     parser.add_argument('--project-fetch-submodules', action="append", default=[], help="fetch submodules for the specified project path")
     parser.add_argument('--include-prefix', action="append", default=[], help="only include paths if they start with the specified prefix")
+    parser.add_argument('--exclude-path', action="append", default=[], help="paths to exclude from fetching")
     parser.add_argument('url', help="manifest URL")
     parser.add_argument('ref', help="manifest ref")
     parser.add_argument('oldrepojson', nargs='*', help="any older repo json files to use for cached sha256s")
@@ -171,7 +184,9 @@ def main():
                    override_project_revs, force_refresh=args.force,
                    mirrors=mirrors,
                    project_fetch_submodules=args.project_fetch_submodules,
-                   include_prefix=args.include_prefix
+                   override_tag=args.override_tag,
+                   include_prefix=args.include_prefix,
+                   exclude_path=args.exclude_path,
                    )
 
 if __name__ == "__main__":