+++ modules/11/default.nix
@@ -15,7 +15,7 @@ mkIf (config.androidVersion == 11) {
   source.dirs."build/make" = {
     patches = [
       ./build_make/0001-Readonly-source-fix.patch
-    ] ++ (optional (config.flavor != "lineageos")
+    ] ++ (optional (!(elem config.flavor [ "grapheneos" "lineageos" ]))
       (pkgs.substituteAll {
         src = ./build_make/0002-Partition-size-fix.patch;
         inherit (pkgs) coreutils;