+++ modules/pixel/default.nix
@@ -28,7 +28,7 @@ let
     coral = { family = "coral"; name = "Pixel 4 XL"; };
     flame = { family = "coral"; name = "Pixel 4"; };
     sunfish = { family = "sunfish"; name = "Pixel 4a"; };
-    bramble = { family = "bramble"; name = "Pixel 4a (5G)"; };
+    bramble = { family = "redfin"; name = "Pixel 4a (5G)"; };
     redfin = { family = "redfin"; name = "Pixel 5"; };
   };
 
@@ -48,8 +48,7 @@ mkMerge [
     deviceDisplayName = mkDefault (deviceMap.${config.device}.name or config.device);
     arch = mkDefault "arm64";
 
-    kernel.name = mkIf (config.deviceFamily == "taimen" || config.deviceFamily == "muskie") (mkDefault "wahoo");
-    kernel.configName = mkDefault config.deviceFamily;
+    kernel.configName = mkOptionDefault config.deviceFamily;
     apv.img = mkIf config.apv.enable (mkDefault (fetchItem imgList));
     apv.ota = mkIf config.apv.enable (mkDefault (fetchItem otaList));
 
@@ -81,6 +80,7 @@ mkMerge [
       "arch/arm64/boot/Image.lz4-dtb"
       "arch/arm64/boot/dtbo.img"
     ];
+    kernel.name = mkDefault "wahoo";
   })
   (mkIf (config.deviceFamily == "crosshatch") {
     signing.avb.mode = "vbmeta_chained";
@@ -129,7 +129,7 @@ mkMerge [
       "arch/arm64/boot/dts/google/qcom-base/sm8150.dtb"
       "arch/arm64/boot/dts/google/qcom-base/sm8150-v2.dtb"
     ];
-    kernel.configName = "floral"; # coral + flame
+    kernel.configName = mkDefault "floral"; # coral + flame
     kernel.linker = "lld";
   })
   (mkIf (config.deviceFamily == "sunfish") {
@@ -139,6 +139,17 @@ mkMerge [
       "arch/arm64/boot/dtbo.img"
       "arch/arm64/boot/dts/google/qcom-base/sdmmagpie.dtb"
     ];
+    kernel.linker = "lld";
+  })
+  (mkIf (config.deviceFamily == "redfin") {
+    signing.avb.mode = "vbmeta_chained_v2";
+    kernel.buildProductFilenames = [
+      "arch/arm64/boot/Image.lz4"
+      "arch/arm64/boot/dtbo.img"
+      "arch/arm64/boot/dts/google/qcom-base/lito.dtb"
+    ];
+    kernel.name = mkDefault "redbull";
+    kernel.configName = mkDefault "redbull"; # redfin + bramble
     kernel.linker = "lld";
   })
 ]